requirements.txt
# OpenAI API for GPT-4 and Codex integration
openai>=0.27.0

# LangChain for prompt chaining and workflow management
langchain>=0.0.148

# PyYAML for configuration file parsing
pyyaml>=6.0

# Transformers for CodeBERT (and other Hugging Face models)
transformers>=4.26.0

# PyTorch as a backend for Transformer models
torch>=1.12.0

# LlamaIndex for repository indexing and retrieval-augmented generation
llama-index>=0.5.0

# PyTest for unit testing
pytest>=7.2.0
